WIP: T02 -- Split web_helpers.py into domain files
