+++ tests/integration/telemetry/test_global_jobs_and_service_info.py
@@ -20,7 +20,7 @@ from hassette.exceptions import TelemetryUnavailableError
 from hassette.scheduler.triggers import Every
 from hassette.schemas.domain_models import ServiceInfo, SystemStatus
 from hassette.schemas.job_models import JobSummary
-from hassette.test_utils.web_helpers import make_job_summary, make_real_job
+from hassette.test_utils.web_job_helpers import make_job_summary, make_real_job
 from hassette.test_utils.web_mocks import create_hassette_stub, create_mock_runtime_query_service
 from hassette.types.enums import ResourceRole, ResourceStatus
 from hassette.web.app import create_fastapi_app
