+++ tests/integration/web_api/test_endpoints.py
@@ -8,7 +8,7 @@ import pytest
 
 from hassette.exceptions import TelemetryUnavailableError
 from hassette.schemas.listener_models import ListenerSummary
-from hassette.test_utils.web_helpers import make_manifest
+from hassette.test_utils.web_manifest_helpers import make_manifest
 from hassette.web.config_view import MASK_SENTINEL
 
 from .conftest import make_log_record
