+++ tests/unit/cli/test_commands_listener.py
@@ -11,7 +11,7 @@ from hassette.cli.commands.listener import (
     cmd_listener,
 )
 from hassette.cli.context import CLIContext
-from hassette.test_utils.web_helpers import make_execution, make_listener_with_summary
+from hassette.test_utils.web_telemetry_helpers import make_execution, make_listener_with_summary
 from tests.unit.cli.conftest import (
     SINCE_EPOCH,
     CLIClientFactory,
