+++ tests/unit/web/test_mappers.py
@@ -5,7 +5,7 @@ from hassette.schemas.domain_models import SystemStatus
 from hassette.schemas.listener_models import ListenerSummary
 from hassette.schemas.live_counts import LiveCounts
 from hassette.test_utils.config import TEST_SOURCE_LOCATION
-from hassette.test_utils.web_helpers import make_manifest
+from hassette.test_utils.web_manifest_helpers import make_manifest
 from hassette.types.enums import ResourceStatus
 from hassette.web.mappers import (
     app_manifest_list_response_from,
